Add a Dockerfile
This adds a `Dockerfile` to build a suitable application container image
and an initial `.dockerignore` file to keep the build context clean.

This uses a multi-stage build to keep the final container as small as
possible. It doesn't currently have an entrypoint, this can be set in
the compose file for now.