script/build: improve image tagging
This ensures that as well as being tagged `mysocietyorg/caps:latest` any
images built are also tagged with the relevant short git commit hash.

This means that as the `latest` tag is replaced the older images will
still have a tag in the `mysocietyorg/caps` namespace and that the git
commit they were built from will be easily visible in the output from
`docker image` rather than just via something like the following:

`docker inspect --format='{{.Config.Labels.revision}}' <IMAGE ID>`

Which is less intuitive and relies on knowing the image ID in question.