Allow location search by council name, not just postcode
Postcode search inputs on homepage and location search results page
now display an autocomplete list of council names when you type
three or more characters, if JavaScript is available.

Selecting a council from the list immediately navigates to that
council’s page, rather than submitting the form (since the form can
only handle postcodes right now).

If the postcode search input is inside an interstitial-triggering form
(which it is, right now), the interstitial-triggering logic is run
before the page location is changed.

Fixes #88.