Move and rename the production.yml file
Due to a change in the way we manage and deploy the live instance of
this service, the old `production.yml` file is being made more clearly
an example of what a production Docker Swarm stack file might look like.

This commit also clears up some other references to this file and
removes the Docker Swarm cases from `script/server` as this script is
not used to manage the production service.

The local `script/build` will now pull images listed in the development
file rather than the production file.