fix issues with relative paths imports in scss for other apps styles
Importing other apps styles is handled differently in live and dev
environments as the dev environment processes the non collected file and
live processes the collected file.

There is an existing ticket for this: https://github.com/jazzband/django-pipeline/issues/749
which has a PR. To fix this we've forked the fixed version and are using
that directly from github rather than pulling in an official package.