Open Graph meta tags, customised via views
As part of this, I’ve moved the definition of the page title from the
`title` block in the templates, to the get_context_data method in the
views. This allows the page title to be used twice in `base.html`, once
in the `<title>` element, and once in `<meta property="og:title">`.

Not sure how best to handle custom open graph image paths for different
pages/views. I’ve hard-coded the domain name for now, but there’s
probably a better way of doing this.