Swap DOM order of council detail sidebar and main content
Sidebar now appears *after* main content in the DOM, improving the UX
on narrow screens. When there’s room for the sidebar to be displayed
in a column alongside the main content, we visually swap the order,
so it appears in a column to the left of the main content, as before.

Part of #157.