add a command to enable reloading the solr schema
The docker scripts won't replace a changed schema so we need a script to
do that for us. This copies a new schema in to the correct place and
then issues an API command to reload the core. The is probably not safe
if you are changing an existing field, but is fine for adding a new one.

You then need to re-index to pick up the changes.