Solr: Fix image in development to solr:6.5
It appears that arm64/v8 support has been removed for older versions of
solr (5 and 6) going forward. This poses an issue developing on M1 Apple
computers, so for now this commit pins the version to one with an image
that includes the necessary arch.

Solr 6 is old, we should probably look at updating but this may depend
on whether support for newer versions has been added to Django Haystack.

See: https://github.com/docker-library/official-images/pull/5973/files