change how we store sections scores, and weighted score types
Weighted scores are always a percentage so use float fields for these.

Also move to storing the max score on the plan section score as the
number of questions varies depending on the council type so makes more
sense to store the max against the section score rather than the
section.