update plan score import in line with data changes
Hard code the sections because they are fixed so makes sense to create
them from a fixed list.

Weighted section scores and non weighted scores are in separate files so
split the loading of those. Similarly the overall score is now in a file
so get it from that rather than calculating it.

Also, all weighted scores are as percentages in data so no longer need
to calculate those. But do turn then from 0.nnxx to nn.xxx because we
only ever want the latter on the site.