Rename "answers" view to "council" for simplicity
I initially just wanted to change the URL for a council page
from /councils/<slug>/answers to just /councils/<slug>/, but
then I realised I may as well remove the "answers" references
in the url name and view as well, since they only confuse
matters. This view is a view of a council. Yes it contains
their scores (answers, at a stretch) but it contains lots of
other stuff besides.