/location endpoint for finding councils by postcode
Fixes part of #177. The postcode search boxes on the homepage and
in the navbar submit to a new /location endpoint, which either
redirects you to the council covering the given postcode, or display
a disambiguation page with the two/three councils covering that area.

The save code duplication, the scoring LocationResultsView extends a
new BaseLocationResultsView from the caps project, meaning we get all
of the postcode lookup stuff from caps, for free.