Styling for plan source URLs on council page
Part of #186.

The source URLs for a council’s marked plans are now contained in a
fold-out `<details>` element, and marked up as a list, for better
accessibility. The `<summary>` element is styled as a button, to match
other controls on the page.

CSS styling on the `.scoring-source-links` automatically truncates them
to the width of the parent, with an ellipsis at the end of the line. To
help get more useful information before the ellipsis, I use a template
filter to remove the URL scheme and www subdomain if present.

(Since we already had a `url_human` template filter, which annoyingly
everything *after* the domain name as well as everything before it, I
renamed that to `domain_human` and made a new `url_human` filter which
*does* include the full URL.)