use manifest storage for static files for cache busting
We also need to move the location of the static files config as
otherwise debug is always False and not affected by the config import.

We're not compressing the JS at the moment, partly because it currently
disagrees with the compressor, only shuffling the location so use the
Noop pipeline compressor.