dismiss data feedback popup with no download if click background
Stops downloading if a user clicks outside the dialog to dismiss it by
setting a property on the original event and using that to check if we
should blank the download field.

Can't do this by stopping propagation as the hide.bs.modal event is
further up the chain than the click event on the dialog background.