[CAPE] Styling and JavaScript for council list advanced filters
Since there now more filters on the council list page, the filters are
displayed in a flexbox grid at the top of the page.

When JavaScript is available, the less frequently used filters are
hidden (and can be toggled with a button at the bottom of the form),
and a live sorting dropdown is revealed at the top of the council list
itself (hiding the sorting dropdown in the form).

Council list rows include new `data-sortable` attributes to enable the
live sorting behaviour.

This commit also introduces new `.d-js-*` utility classes, that act like
Bootstrap’s `d-print-*` classes, but trigger when JavaScript is enabled.