Configure PERSONAL_ACCESS_TOKEN in the loca settings file
Rather than call out to the environment in multiple places, let's set
the PERSONAL_ACCESS_TOKEN variable in our settings file then reference
that from the code itself.

This will make it easier to template the settings file and provide
settings in ways other than through the environment, should we need to.