Add a script to update Solr config and schema when running locally
Ideally, solrconfig.xml should be seeded when the core is initialised as
there are some settings that won't take effect on a reload. See the Solr
documentation for more details.

Also, this script will need to be run as the root user to ensure that
it can write into the Solr directories and run the necessary `chown`.