Rebrand as less of a CEUK project, more of a mySociety project
CEUK is focussing more on the Scorecards work, so it makes less and
less sense to have CAPE branded as part of their site.

This commit removes the CEUK header from the site, adds both mySociety
and CEUK logos to the social sharing image, adds a clearer description
of both organisations to the About page, and rearranges the main menu
as a result (splitting out Data and About once more).

Also moves all data-related stuff from /about to /about/data.