[CAPE] make council table sort in place with javascript
Add a change handler to the sort order dropdown, plus some data
attributes on table cells to make sorting easier. Special handling to
always list things without a date at the end of the table when sorting
by date.

The JS sorting is only displayed if JS is present otherwise the sorting
is displayed in the default filters, which is hidden by JS.

Also, use CSS to show/hide the additional filtering elements so this
works fine without JS.

This also changes the filter form to render the individual elements so
we can control where they go. We also display the sort element twice to
make sure that submitting the filtering form results in the JS sort
displaying the correct value.