[CAPE] Styling and JavaScript for council list advanced filters
Filters are now displayed in a flexbox grid at the top of the council
list page. When JavaScript is available, the less frequently used
filters are hidden (and can be toggled with a button at the bottom of
the form), and a live sorting dropdown is revealed at the top of the
council list itself (hiding the sorting dropdown in the form).

This commit also introduces new `.d-js-*` utility classes, that act like
Bootstrap’s `d-print-*` classes, but trigger when JavaScript is enabled.