[CAPE] Fix colour bug in equilateral-triangle mixin
The equilateral-triangle mixin was set to use the parent’s
_border color_ as its own colour, not the parent’s _text colour_,
which is what was intended.

This meant that, when using the equilateral-triangle mixin inside a
`.btn`, say, the triangle would match the colour of the button border
and not the colour of the text inside the button.

Now, the triangles will match the colour of the text around them, and
will play nicely with Bootstrap `.text-*` colour utility functions.