[Scorecards] Fix methodology 2023 accordion bugs
Created a new `.toggle-section` CSS component for styling the section
accordion, so that we can keep using the old `.accordion-wrapper`
component on the more complicated question.html page.

The toggle interactivity is handed through a new `.js-hidden-toggle`
component, which uses `aria-controls`, `aria-expanded`, and `hidden`
(rather than CSS classes) to toggle visibility.

The `plus-*.svg` icons were really messy, so I tidied them up too.

And the styling for the accordion on the How To Use The Scorecards page
was broken too, so I fixed that.