[CAPE] Get site url from django-hosts, for templates
Enables us to keep the site URL out of templates, making it easier to
keep absolute links up to date if we ever change URL (which we’re likely
to do as part of https://github.com/mysociety/caps/issues/225).

Thanks to @struan for noticing that the django-hosts package has a nice
easy template tag for handling this!

Fixes #146.