Fix :first-child() / :last-child() CSS typos
29957e57149c1a14bfe6e78f81caf46c2162aa5a introduced a CSS typo in
twins.scss, where the first and last child pseudo-selectors were
followed by empty brackets. This worked at the time I wrote it
(perhaps the version of Sass we were using at that point handled the
mistake gracefully?) but our upcoming scorecards-2023 branches use
a new Sass compiler, which treats this as invalid CSS and causes
django pipeline to raise an exception.

Same goes for the :first-child() in the 2021 scorecards methodology.