Un-inline SVG background images in Sass files
An attempt to bypass an issue we’re having with django-pipeline
where csscompressor is stripping the whitespace out of the
inside of inline SVG markup in CSS files. By telling Bootstrap
to use external SVG files instead, the SVG markup is no longer
in the CSS for csscompressor to mess up.

I also had to do the same for the custom chevrons we use for
table sorting. While I was there, I deleted all the unused
chevron icons that were copied over from the 2022 scorecards.