[Scorecards] Standardise `scoring_group` references
We previously used `authority_type` in a number of places, when really
we were referring to Scorecards scoring groups (eg: "Single Tier"), not
a council’s _actual_ MapIt `authority_type` as the name might imply. We
also sometimes used the term `council_group` for the same thing. Sigh.

This commit is an attempt to disambiguate all those `authority_type`
and `council_group` references, so that now, whenever we’re talking
about a Scorecards scoring group, we call it `scoring_group` and we pass
the group’s full object from Council.SCORING_GROUPS, not just its slug.

This involved a number of fiddly changes. including:

- Removing `caps/includes/authority_type.html` and using
  `scoring/includes/scoring-group-name.html` instead.
- Replacing `context["authority_type"]` with `context["scoring_group"]`
  in all post-2022 views and templates.
- Replacing `council_group` with `scoring_group` in `PlanSection.get_average_scores`
  and `PlanQuestion.get_scores_breakdown`.
- …And while I was there, I also noticed that `PlanQuestion.get_average_scores`
  wasn’t called anywhere, so I removed it.