[Scorecards] More obvious styling for active advanced filters
Rather than having an element that reads out which advanced filters are
active (which we’d already temporarily hidden because we were worried it
was just too much clutter) and a "reset" button, this commit introduces
very obvious styling for advanced filter `<select>` elements when they
have been modified. This should make it clear which filters a user has
selected, and therefore also easier to quickly reset them manually back
to the Any/All option.

The active filters are styled up to match the active buttons above,
for consistency.

I also made the advanced filter select boxes slightly smaller, to
match the size of the Country buttons on the Single Tier scorecard.