[scorecards] handle previous year filter not returning any results
In some cases when getting the section averages for the home page
running the filter on the previous year might not return any results so
we can't do the section comparison. In that case just skip them. This
doesn't break for the total because it just grabs the results directly.