[Scorecards] Major overhaul of homepage table styling and functionality
I had to change quite a bit of stuff in order to allow sorting of the
homepage table by the "year difference" columns. In the process, I took
the opportunity to remove a load of repetition and inconsistency that
had built up over the years.

It is now possible to:

- Sort the table by any score or year difference column
- …and show/hide the year difference columns
- …on mobile and desktop viewports
- …for both 2023 and 2025 homepages

Other improvements I made while getting there:

- Removed mobile-specific Scorecard table markup. Mobile devices now
  use the same <table> as desktop devices, and we use CSS styles and the
  `data-section` attribute to hide/show one set of columns at a time
  for mobile users. This has more than halved the number of DOM elements
  in the page (down to 10400-ish, from 26600-ish).
- Council names are no longer indented, partly to save space on mobile
  devices, and party to avoid odd spacing and wrapping of long names.
- Fixed missing council type label in top left cell.
- Fixed missing council type wording in "no scorecard" error cells.
- Fixed a bug where column headers didn’t have the right section colour
  (in section_link_with_year.html).
- Remove unused home_combined.html template (combined authorities have
  been handled by the main home.html template for a while now).

The table makes use of colspan and rowspan to have some header cells
span multiple children when there are multiple years of data. These
colspan and rowspan attributes are rendered first in the template (based
on whether a previous year’s data is available to show – if not, there’s
no need for spanning) and then—for desktop users only—updated when you
enable/disable the year difference functionality. This means you _can_
get the table into a slightly broken-looking state by disabling the year
difference columns in desktop mode, then narrowing your window to mobile
width, which now lacks the colspan/rowspan values it expects. This is an
annoyingly tricky edge-case to avoid, since you can’t set col/rowspan
attributes via CSS, so would need to monitor for media query breakpoints
in JavaScript instead. I’ve decided the extra complexity of a workaround
for this isn’t worth it, as very users will switch between desktop and
mobile viewport widths, and of them, only those who have _also_ disabled
the year difference columns will see anything wrong. And even if they
see it, it’s only the table headers that look a little off-by-one.