[Scorecards] Add Homepage-style filter buttons to Question page
Switching to the Homepage-style filter buttons will make it easier for
us to incorporate Homepage-style filtering by nation/country/etc to the
Question pages (#769).

Now that the “homepage” filter is being reused on three page types
(home, nation, question) this was a good opportunity to generalise
some of the code, including:

- rename `#home-page-main-filter` to `#main-filter`
- rename `ajaxLoadCouncilTypeScorecard` to `ajaxLoadPageFragments`
- separate `.loading-shimmer` from `ajaxLoadPageFragments` so that
  specific child elements can _appear_ to be reloaded, while the
  JavaScript simply reloads their common parent.

I also took the opportunity to tidy up the typography and vertical
rhythm of the Question page sections/headings.