[WIP] Filtered council counts in Question page statistics
Since the table of councils/scores on the Question page can now be
filtered by attributes like nation/region/imd/etc, it makes intuitive
sense that the statistics _above_ the table should _also_ respect those
filters.

The rising/falling cards already work with the filters, which is nice.
But the breakdown of total councils with each number of points does not.

A challenge here is that the points breakdowns include a comparison
with the previous year’s data. But filters are year-specific. What does
it mean to compare “Labour-controlled English Single Tiers” from 2025
and 2023?

I had a bit of a debate with Claude, and it suggested that we make this
change – comparing the stats of _the same specific councils_ across the
years, rather than trying to apply the same _filters_ across the years.
This way, you’ll see how many of the matching councils got each number
of points in 2025, and how many points they all got in 2023 too.

Seems fairly intuitive to me? But I’d appreciate a second opinion!

I am also a little bit “yuck” over how it’s re-implementing something
new, instead of re-using `get_scores_breakdown`, but maybe that’s just
un-avoidable.